requirements.txt
Django==5.2.4
djangorestframework==3.16.0
faiss-cpu==1.11.0
gunicorn==23.0.0
langchain-community==0.3.27
langchain-core==0.3.68
langchain-huggingface==0.3.0
langchain==0.3.26
python-decouple==3.8
sentence-transformers==5.0.0
torch==2.7.1
transformers==4.53.2
